# Don't copy the array for incoming msgpack reads